chrome_ec: disable ccd when we use servo v4
We need to disable ccd even if we are using servo micro. The servo v4
captive cable is used for the data connection between servo v4 and the
DUT even if we are using servo micro instead of ccd. We want to make
sure that cable does not appear as a debug accessory, because that may
cause cr50 to enable ccd even if we are not using it. This change
disables the debug mode when ccd_disable is called and we are using
servo v4 to run the test.

BUG=none
BRANCH=none
TEST=plug the servo v4 captive cable into the cr50 ccd port and run
firmware_Cr50USB

Change-Id: I20a113b0864962d2b5e1735e14024a213f3be4b8
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/513618
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>