Acommodate whitespace or lack of whitespace before ec version.
This is a quickfix to address failing tests.  Likely a longer term fix
would be to modify the tools that are relied on to gather version
information to return something structed (e.g. a protocol buffer)
instead of maintaining logic to parse the presentation layer of those
tools (which is subject to change).

BUG=chromium:726119
TEST=took a logged 'chromeos-firmwareupdate -V' and ran through the
splitting logic in an interpreter.  Also tweaked the logged version to
test the case when there was a space prior to the version.

Change-Id: I7b8aa5de800ffc1921f95f1e5239d4a50ebbd4e8
Reviewed-on: https://chromium-review.googlesource.com/514369
Commit-Ready: Kevin Shelton <kmshelton@chromium.org>
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>