[autotest] Respect exclusive --in-lab or --host-attributes arguments
These two arguments to autoserv are not intended to be used together.
The only user of --host-attributes are test_that / test_droid that use
it for tests run outside the context of the lab via
test_runner_utils.perform_run_from_autotest_root.

This CL enforces this mutual exclusion to make way for a future CL that
depends on this property.

+ As we begin to use HostInfoStore for storing the attribute
information, copy the host_attributes into the store.

BUG=chromium:678430
TEST=(1) test_that / test_droid still works.
     (2) in-lab hwtests via moblab.
CQ-DEPEND=CL:519569

Change-Id: I0dcb3096a4797143df796855bf3babc155e76c67
Reviewed-on: https://chromium-review.googlesource.com/517771
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>