[autotest] Use process pool to read pidfiles in drone refresh
We spend a lot of time reading each pidfile corresponding to the running
processes on drones. Use a multiprocessing.Pool to speed this up.

For now, we hide the use of multiprocessing.Pool behind a global_config
flag. This allows us to gently roll this out to shards observing the
impact.

BUG=chromium:718181
TEST=[1] unittests.
     [2] Run a test on local autotest instance.
     [3] moblab_RunSuite.

Change-Id: I08333453711a07ea3d7b805b44fda8e032a535eb
Reviewed-on: https://chromium-review.googlesource.com/525075
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>