[autotest] Add support to merge directory summaries from multiple result collect
This change adds the main logic to merge the directory summaries from the
multiple result collections during the test. During the merge, it also calculate
the total size of result files transfered from the test device.
At the end, the summaries are merged with all the files in the test result
folder on the drone, which will be used to compile the final view of test
results.

BUG=chromium:716218
TEST=unittest

Change-Id: I3ff326a046497f752125ed679e7c4f72bdddc5fa
Reviewed-on: https://chromium-review.googlesource.com/526274
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Dan Shi <dshi@google.com>