[autotest] Hotfix stage_server_side_package undefined variable
This function needs to be cleaned, but I'm not going to do it right now.
This CL will make stage_server_side_package not fail in the purportedly
supported case of a DUT that has a build assigned but no job_repo_url.
This may still fail later in the flow somewhere.

The fact that nobody noticed this probably means we don't really support
this case.

BUG=chromium:730431
TEST=None.

Change-Id: I984caf4d70bd5b58f1d39678c8e62a0e7c19f397
Reviewed-on: https://chromium-review.googlesource.com/527600
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>