Adding reference of CROS_TH version for proper labeling
CrOS tree hugger is ARC CQ which has a different cros version label.
To ensure that the DUT gets labeled with the right prefix CROS_TH label by
provisioning, adding the prefix reference for the new label in provision and
host code. The actual test changes are coming in a follow up CL.

This change is to ensure that I am not missing any other references.

TEST=None
BUG=b:26895109

Change-Id: I491b72d9544d5398ac67a171838e7f0b1caf12e3
Reviewed-on: https://chromium-review.googlesource.com/528628
Commit-Ready: Rohit Makasana <rohitbm@chromium.org>
Tested-by: Rohit Makasana <rohitbm@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>