[autotest] Add ShadowingStore.
This is a transitional CachingHostInfoStore that allows us to continue
to use the AfeStore in autoserv, but also shadow it with a FileStore.
This will help us uncover any divergence in the two stores, which may
arise as a result of AfeStore getting updated out-of-band, invisible to
the FileStore.

BUG=chromium:678430
TEST=(new) unittests.

Change-Id: I3f473cc93411051a188497ef10124ce5ff59421d
Reviewed-on: https://chromium-review.googlesource.com/530126
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>