Have lsbrelease_utils function respect argument default.
The API of the lsbrelease_utils functions says that when the
lsb_release_content is None, we try to infer it ourselves. But the
implementation tries to infer the content even when the argument is
passed in and is ''. This is not ideal because a call from drone may
result in an empty lsb_release_content being returned from a DUT. A
subsequent call to lsbrelease_utils functions will incorrectly check the
/etc/lsb-release on the drone (if present).

BUG=b:28453441
TEST=None.
     - Spot checked manually that none of the callers would be impacted
       by this change. This is especially difficult because all these
       functions are imported  by another utils module and used from
       there.

Change-Id: I2d6ded6016fb6e373de3969632f02ad9bc027cd5
Reviewed-on: https://chromium-review.googlesource.com/536414
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>