[autotest] Use ShadowingHostInfoStore in autoserv
Instead of using an AfeStore to obtain host information from the AFE,
this CL switches to a ShadowingHostInfoStore. The information still
comes from the Afe, but is shadowed by a file on disk. Any differences
between the local store and Afe are reported to monarch.

BUG=chromium:678430
TEST=Run a job on local autotest instance with and without ssp.

Change-Id: Ifa3da3c3449ac08bfc53dcef0c3e871b36b69602
Reviewed-on: https://chromium-review.googlesource.com/537976
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>