network_WiFi_SuspendStress: don't look for servo for non-servo tests
We stopped requiring servo for these, so stop requesting one. Otherwise,
we might fail this test when run without servo (we don't always catch
this with 'IOError' it seems).

The only downside I see here is that without even trying to request
servo, we now don't get to reset the device if it hangs.

BUG=chromium:707988
TEST=network_WiFi_SuspendStress.xxx variation, with no servo

Change-Id: Ibf7d256e40f5f4c2831c503e35a29ad511e1992c
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/540195
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>