[autotest] Remove leftover servo references from autoupdate_EndToEndTest.
This test used to feature a bunch of references to servo. This work was
removed but there are a few things left over in the test.

I will remove them now to make the test easier to follow.

More details in CL:302542 and CL:302871

BUG=chromium:709710
TEST=test_that <ip> autoupdate_EndToEndTest

Change-Id: I97cde59286335017715a9be779bd1a8c43205210
Reviewed-on: https://chromium-review.googlesource.com/540187
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>