firmware_Cr50Update: check mainfw_type after update
I was running firmware_Cr50Update overnight and when I came back in the
morning I saw that the ssh command timed out. I couldn't tell if it is
because the device booted into recovery or the command timed out for
some other reason. I am changing the test to check that the main fw type
is normal after the update. Now if the device boots into recovery, we
should be able to tell.

BUG=none
BRANCH=none
TEST=run firmware_Cr50Update and see if I can detect the device booting
into recovery mode.

Change-Id: I4f8eecda73128a01af3aff6b58fe808b52993ca5
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/541597
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>