firmware_Cr50InvalidateRW: change order of tests
Clearing the tpm owner will prevent the rw header from being invalidated
until the second login. Before firmware_Cr50InvalidateRW would reset
cryptohome first and verify its login. If for some reason the tpm owner
was cleared right before running this test, then ChromeOS would not try
to invalidate the header until the second login. This change moves
clearing the tpm owner to the front of the test, because it resets the
most state, and makes sure that we know what state the device is in.

BUG=none
BRANCH=none
TEST=run the test after clearing the tpm owner and make sure it succeeds

Change-Id: Ifbfbc13bd07eadfdc4545d9254c6e1cc20293181
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/541598
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>