[Autotest] Increase max allowed time for ChromeVox intro message.
The accessibility_ChromeVoxSound test verifies that ChromeVox
makes noise when it is enabled or after certain actions are done.

The welcome message for ChromeVox has varied in length over the
past milestones, as it is a common place for alerts and tips
about new features.  Since we don't actually care how long
this message is, increase the maximum allowed time to be 20 seconds.

TEST=ran it
BUG=chromium:553713

Change-Id: Ief0fb97c9f25525c33035b1f8df3d62b08cc0a3f
Reviewed-on: https://chromium-review.googlesource.com/540832
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>