bluetooth: fix adapter reset method
In the beginning of every autotest, we need to reset the adapter
to make sure that it is in an initially clean state. Instead of
restarting the bluetoothd, it is better to remove device objects
through dbus methods. Otherwise, the unhappy chrome browser may
disable the bluetooth adapter in a random manner.

BUG=chromium:708915
TEST=Run the test
(cr) $ test_that --args "chameleon_host=${CHAMELEON_IP}"
       ${CHROMEBOOK_IP} bluetooth_AdapterPairing.mouse

Change-Id: Ic0ec1f5626b2e33ea46dfbc7125168f3d7da0d5f
Reviewed-on: https://chromium-review.googlesource.com/571250
Commit-Ready: Ruchi Jahagirdar <rjahagir@chromium.org>
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Ruchi Jahagirdar <rjahagir@chromium.org>