autotest: add test verifying cr50 boardid management
This change adds a test to verify cr50 boardid behavior. It tests
setting and getting a boardid on a board id locked image including
verifying that cr50 will reject an incorrect boardid. It also verifies
that if a mismatched boardid has been set, and we update to the board id
locked image, cr50 will rollback to the old image.

Prod images cannot erase the board id, so this test requires a
CR50_DEV=1 image to run. It can be stored locally or in google storage,
but it needs to have the ability to get/set boardid and rollback.

BUG=b:62876435
BRANCH=none
TEST=none

Change-Id: Ib1ee5858a8b6ccc4868859caa0b69e374d659c80
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/558652
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>