autotest: Add chrome-informational test suite
Move desktopui_MusLogin and desktopui_MashLogin to the new suite.
These tests need to run on hardware, but are for features in Chrome
that are still under development behind flags. We don't want to
block Chrome PFQ uprev if they fail, just inform the gardener.

This gives us an alternative to bvt-perbuild that we can run
continuously against ToT chrome to provide narrower chrome
revision ranges for regressions.

Remove hacky board-specific control files for these tests. They
were an attempt to put the tests in bvt-cq for just a couple of
boards, but if we introduce a flaky regression in these features
we don't want to block the CQ.

BUG=chromium:737240
TEST=presubmits
Change-Id: I97af6d24793c2fed529b3cc799960c07506228ac
Reviewed-on: https://chromium-review.googlesource.com/550863
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Tested-by: James Cook <jamescook@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>