[autotest] Remove more dead AU code.
server/autoupdate_utils is not used by anything and it is lives
confusingly close to server/cros/autoupdate_utils that is used.

autoupdate_ParameterizedJob is dead code. I searched the code base and
found a function that looks up a db for that name and returns True if it
was found, else false. So I'm assuming that the function always returns
False and so it can be removed along with its single caller.

BUG=chromium:709710
TEST=scheduler_models_unittest

Change-Id: I609914b5b8dc4be657a57cd60d4b3dcb1b3562f0
Reviewed-on: https://chromium-review.googlesource.com/550568
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>