cr50_stress_experimental: add test verifying console
To catch basic mistakes with locked/unlocked commands, the pinmux, or
gpio config add a test to verify the console output for help, pinmux,
and gpiocfg. This test will confirm that the current output has not
changed. Every time one of these commands changes, we will need to
update gpiocfg, help, or pinmux. The test does not care about order, so
we can easily add new commands to help or new lines to pinmux.

This test is mostly to make sure that we notice the changes to the
important console commands.

There are different gpiocfg and pinmux configurations for boards with
different properties. The command files have different suffixes based on
which board properties they are used for. I added the files for the main
boards. In the future, we may need to add more.

BUG=b:35647854
BRANCH=none
TEST=Run on 20 and 21 with ccd enabled and disabled

Change-Id: I99b10c911909426eff3cec37ccdda162b23af13b
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/550877
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>