[autotest] Make sysinfo logging more robust
We ignore errors in sysinfo log creation / collection. But we bailed on
the first error encountered so that the rest of the requested loggables
were never run. Since we're anyway ignoring errors, ignore them for each
loggable as well so that we get more information on failures.

BUG=chromium:728290
TEST=(1) new unittest.
     (2) Run a client test and notice some error messages about (current
         failing) log collection, and verify that other loggables are
         still executed.

Change-Id: Ie09df2d0510f8a225a5d6cdde0f4dc59d55e3c9c
Reviewed-on: https://chromium-review.googlesource.com/558202
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>