ui: Sort by created_on when filtering created_on
The JobStartingDatetimeFilter puts a constraint on created_on (even if not
specificed, because if blank, it still uses a setStartTimeToPlaceHolderValue()
regardless). When the query has uses ORDER BY id, the InnoDB query optimizer
uses the primary key index depsite having a constraint on created_on which may
cause it to scan through an unbounded number of rows before actually finding
rows that satisfy the created_on constraint. By switching to ORDER BY
created_on, the query optimizer switches to using the created_on index, as it
should.

BUG=chromium:685424
TEST=None

Change-Id: Id38e054170264d8697cecc5d8607c8f5ee61f761
Reviewed-on: https://chromium-review.googlesource.com/558669
Commit-Ready: Paul Hobbs <phobbs@google.com>
Tested-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>