[autotest] Delay chromite imports in site_utils
The chromite imports in site_utils are specific to some specialized
use-cases, but cost all users 2-3 seconds total during module import.
This was a bottleneck in the scheduler tick time.

The correct fix is to break apart server/site_utils since it does way
too much. The hack is to delay the imports.

BUG=chromium:739466
TEST=None

Change-Id: Ia28481414b5db8b413d70839076ddcc50d747f6e
Reviewed-on: https://chromium-review.googlesource.com/559831
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>