Add autozoom sanity test.
Adding autozoom sanity test that enrolls the device into CFM and runs a
meetings session for a pre-defined length. After the session ends, it
examines the video sent resolution in JMI logs and outputs it into a
json file to be consumed by Chrome Perf Dashboard.

TODO's for this test include adding logic to examine CFM sent resolution
and taking appropriate action if resolution is not in the expected
range. Another TODO is to implement check_verify_rtanalytics_logs
method.

BUG=chromium:739580
TEST=Tested on local setup.

Change-Id: I08d1d1c06ac6cd6ce5521920fd98f8a568831c50
Reviewed-on: https://chromium-review.googlesource.com/560627
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Reviewed-by: Dirk Kaeding <dkaeding@google.com>