[autotest] Add new script to prefer wifi network interface vs ethernet.
This is an adapted version of the code in
src/aosp/system/connectivity/shill/test-scripts that is adapted to work
on moblab.

It switches the priority of the network connections so wifi connection
is the first connection, however all wifi traffic is still routed
through the ethernet connection.

This is required for certain CTS tests to pass.

BUG=chromium:739582
TEST=Manual testing on moblab.

Change-Id: I552403d92c739a2bb2e830b73a2c53a42390e4f6
Reviewed-on: https://chromium-review.googlesource.com/560583
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>