Implements a WebRTC audio loopback test that computes RMS
The test start a loopback call with fake audio, runs it for 10 seconds
and creates a recording using cras_utils. The recording is then analyzed
and RMS is reported to the perf dashboard.

The test runs for now only on bluestreak devices since they are all
equiped with a speaker (mandatory to be able to compute RMS).
In the future though we should extend this to run on other platforms
(e.g. Chromebooks or older CfMs) as well.

Also updates the perf_dashboard config.

TEST=manually
BUG=chromium:698147

Change-Id: Ie55cdc4ae6c5fb23732f6194aca1ea4dd3216116
Reviewed-on: https://chromium-review.googlesource.com/563408
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Denis Tosic <dtosic@chromium.org>