Add performance monitoring back to the AU test.
In CL:544536 the performance monitoring was removed because the device
was restarted before it could retrieve the running test results.

This CL adds the performance monitoring back to the test.
It copies the script to the DUT. During the update the chromite
auto_updater will kick it off and save the results when done.

This CL then pulls the file back from the DUT and records the metrics.
This can be submitted before or after the chromite CL.

BUG=chromium:709710
TEST=autoupdate_EndToEnd passes

Change-Id: I4cb59e30ac68c1e97ca74136f51bac3f76be050f
Reviewed-on: https://chromium-review.googlesource.com/564340
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>