Updated enterprise_CFM_USBPeripheralHotplugDetect test to detect crashes.
Check for crash function looks for crash files which ends with or
contains '.meta', '.kcrash' as part of their name and reports
them as crashes. crashes_list contains the information as when
the crash file is detected. This test uploads the total crashes
count to perf_dashboard for metrics.

BUG=chromium:709134
TEST=Tested on Panther

	modified:   enterprise_CFM_USBPeripheralHotplugDetect.py
        modified:   ../../../tko/perf_upload/perf_dashboard_config.json

Change-Id: Ic380355f867bb6225060f9c0e069ba0c4e78762e
Reviewed-on: https://chromium-review.googlesource.com/564283
Commit-Ready: Dinesh Kumar Sunkara <dsunkara@chromium.org>
Tested-by: Dinesh Kumar Sunkara <dsunkara@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Dinesh Kumar Sunkara <dsunkara@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>