[autotest] Stage stateful to the same location as rootfs payloads.
The autoupdate_EndToEndTest makes use of the devserver cros_au RPC that
expects the rootfs and stateful payloads to be staged within the same
static folder on the devserver.

When run in the lab the test was staging the stateful payload for the
target update in a different location because it was using the job_repo_url
to figure out where stateful is.

This CL makes it look in the same folder for the stateful.

BUG=chromium:741779
TEST=autoupdate_EndToEndTest passes

Change-Id: I2c8a98d9c532eaef0d4fd5566f22c199fe842c0b
Reviewed-on: https://chromium-review.googlesource.com/568390
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>