security_ASLR: More robust PID calculation.
security_ASLR test wants to find the PID of a single process by name
that is also a child of a named parent process. Unfortunately, the
selection of the PPID to filter the process list is currently
somewhat arbitrary - recently we added a "use the lowest PID"
heuristic to work around the fact that there could be multiple
'init' processes.

This change instead makes the code find all candidate PPIDs, then
filter the ps output to match any PPID, and only succeed if it
finds exactly one PID/PPID pair that match the names.

BUG=chromium:741110
TEST=test_that security_ASLR # passes with multiple 'init' processes
TEST=test_that security_ASLR # catches missing pid with \
                               update_engine stopped.

Change-Id: I8185248f2dc1b80f706f1f97c11a0ff7a4fd35fd
Reviewed-on: https://chromium-review.googlesource.com/568905
Commit-Ready: Josh Horwich <jhorwich@chromium.org>
Tested-by: Josh Horwich <jhorwich@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>