Adds new Google Meet perf test for CFM
Extends the existing perf test with support for joining a meeting session.
The meeting session is always the same one (for now) and will be
prepopulated with participants.

The type of test is controlled through the new is_meeting param (defaults to False).
Updated the perf_dashboard_config to now include the new test.

Also cleans up the code a bit.

BUG=chromium:741644
TEST=on local setup

Change-Id: I04edb36e70494b1da1f83cf4b3486ec0c92bf196
Reviewed-on: https://chromium-review.googlesource.com/570046
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Denis Tosic <dtosic@chromium.org>