[autotest] Don't run set_gbb_flags for servo reboot repair.
The ServoRebootRepair class was calling its super-class for the
final acutal reboot call.  That inheritance caused it to invoke the
'set_gbb_flags' script.  However, the 'set_gbb_flags' invocation
was only meant to run on CrOS DUTs, not servo.

This moves the problem invocation to a new subclass that is specific
to CrOS DUTs.

BUG=chromium:735159
TEST=invoke the new repair action from the python CLI

Change-Id: Id12c79b547aa7ee4c9ee40a0344940e243ff7d80
Reviewed-on: https://chromium-review.googlesource.com/572057
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>