network: clean up error message in CheckInterfaceForDestination
Prior to this patch, the error message was an AttributeError
that said there was no interface attribute on NoneType if
NetworkRoutes.getRouteFor returned None. This is a pretty obscure
error message so we should check for None and raise a better one
in that case.

BUG=None
TEST=network_3GSmokeTest

Change-Id: If48aa4b9971d5538f77c3a351280f7dba5295dcc
Reviewed-on: https://chromium-review.googlesource.com/575012
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>