Add model label to autotest DUTs for unibuilds
CBuildBot needs to schedule specific tests in autotest based on the
model.  With unified builds, the board (aka build) is no longer granular
enough for scheduling test coverage.

This adds the model:[modelname] label to test targets in the autotest
farm, which will be used for test scheduling in the future.

BUG=chromium:576363
TEST=no staging environment available

Change-Id: I07633950732a389b3c3d92f801f63000d4ee7326
Reviewed-on: https://chromium-review.googlesource.com/576363
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>