autotest: delete host_history.py, all callers (depends on autotest_es)
Note: this also deprecates a afe rpc named "get_host_history" which
depends indirectly on autotest_es. This is not to be confused with
lib/status_history:get_host_history which is a separate afe+tko-backed
host history fetcher and is still in use.

BUG=chromium:738508
TEST=None

Change-Id: I6c70188ac1ee6c0c8c097c3f58665e4afc4613bf
Reviewed-on: https://chromium-review.googlesource.com/575575
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>