network_WiFi_TDLSPing: retry "initial" ARP lookups
Apparently, a 1 second delay is not really enough time to guarantee a
lookup succeeds. We're not really aiming to test ARP lookup latency
here, so it makes sense to just retry a few times, especially
considering that on many runs, shill is telling us that "ARP lookup is
still in progress".

Also, add a bit of logging in wifi_client for these TDLS operations. We
do something similar for association and ping-test logic, and it makes
for a little easier reading/analysis when these operations are logged in
the autoserv logs, and not just implicitly through (for example) the
device syslog/netlog.

BUG=chromium:746513
TEST=network_WiFi_TDLSPing

Change-Id: Idc9b9505b6d2744aac4ac0b8bd37a4acf6f5fe73
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/578311
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>