Each step in an update chain can only be one event.
During the AU test we look for a set chain of events (e.g download started,
download finished etc). Once upon a time when update_engine was updated we
needed to account for one of two different event codes that it MAY have output.
This was done with CL:311414.

Since we are never running this test on a build earlier that M53 we can
get rid of this code.

The real reason for this change is it will make rewriting the error messaging
simple. The error messaging from this test is often useless. I'll fix
that in an upcoming CL.

BUG=chromium:709710
TEST=autoupdate_endtoendtest works

Change-Id: I6bf655ae1ed90eb0a4a61f9b6e23a6a9473d0091
Reviewed-on: https://chromium-review.googlesource.com/583946
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>