Improve the error messaging for autoupdate_EndToEndTest.
The test verifies that the update went through a set of expected steps.
The update event steps are received from the update_engine hostlog.
We compare the hostlog events against a set of expected events.
This is done AFTER the update is complete.

The old way was to compare the events while the update was in progress
and the current error messages still read like this is the case.

I've updated the error messages for the most common failures to explain
what actually happened:
1. initial update check
2. timeout
3. post reboot update check.

The omaha log is no longer included in the test so I've removed any
mention of that.

I will be refactoring this test a bit soon and will look to improve
the error messaging code then.

BUG=chromium:709710
TEST=autoupdate_endtoendtest with all error conditions

Change-Id: Ie08e4fe7a5b8459611ba073f0778500bb08b0b64
Reviewed-on: https://chromium-review.googlesource.com/585928
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>