autotest: enable power measurement by powercap sysfs
This patch enables x86_64 RAPL power metrics by means of the Linux
kernel powercap driver, such that, it gets rid of microarchitecture
knowledge of userspace MSRs polling; Says that if the OS supports
powercap sysfs, the test creates a list of Powercap instances for
measurement, else it falls back to MSRs polling.

BUG=b:63877056
TEST=run 'test_that --autotest_dir ~/trunk/src/third_party/autotest/files/
${DUT_IP} f:.*power_LoadTest/control.fast', examine test report w/ RAPL
metrics of pkg/pp0/pp1/dram/psys domains; note that the list of doamins
is different from platform to platform.

Change-Id: Ided913fd3de840db2d5faaf2284f8dc9a00ba2bc
Signed-off-by: Harry Pan <harry.pan@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/593507
Reviewed-by: Sumeet R Pawnikar <sumeet.r.pawnikar@intel.com>
Reviewed-by: Todd Broch <tbroch@chromium.org>