[autotest] pass-through all host attribute updates to master
This brings attribute update behaviour on par with label updates. If a
set_host_attribute RPC is made to the shard directly, it passes it to
the master, which updates both its own DB and fans out the RPC back to
the relevant shards.

BUG=chromium:749200
TEST=(sanity) Test run on local AFE; run repair on a test_push shard.

Change-Id: I745e2b29e075d4077957169ffa2412c4f4916ea9
Reviewed-on: https://chromium-review.googlesource.com/587417
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>