Handle missing route in network.CheckInterfaceForDestination.
network.CheckInterfaceForDestination checks if any route to a host goes
through a particular interface. As multiple IP addresses may be resolved
for the host, it's ok if some but not all addresses don't have a route.
This CL modifies network.CheckInterfaceForDestination to fail only if
there is no route to reach the host or if some route goes through an
interface other than the specified one.

BUG=chromium:742043
TEST=Run network_3GSmokeTest on a DUT with cellular modem.

Change-Id: I3884ec139110c8f6559480dd8605f741606bc5cc
Reviewed-on: https://chromium-review.googlesource.com/592404
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>