moblab_RunSuite: Really use board label to infer board.
platform of a host is not the same as the board label. In some cases,
board is available (it comes from the OS image) when platform is not
(depends on certain hardware utilities etc). So, really use board, which
is what was being claimed.

This makes it possible to infer and use this label when both the moblab
and it's DUT are just VMs

BUG=chromium:725620
TEST=(1) Run moblab_RunSuite where both the moblab and the sub-DUT are
         VMs.
     (2) Run moblab_RunSuite on real hardware.

Change-Id: I2ae9a5110216bee6ebe0359cd1f3f2611d408efb
Reviewed-on: https://chromium-review.googlesource.com/595242
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>