Update suite owner and add hotrod-release suite
Automated bug filing is no longer available but emails alerts are still active.
These emails are sent to the 'owner' in the BUG_TEMPLATE dict. In this
case we want these emails to go to chromeos-meetings@chromium.org
(existing group that also receives CFM perf alerts).

Also adding a hotrod-release suite which will run a smaller subset of
tests on each new build in <=tot-1 branch builds. This is to be used for
CrOS build sanity qual for both beta and stable builds.

BUG=chromium:753021
TEST=None

Change-Id: Ib546a7ab20bf0ece9636465c2fff92190f6a1759
Reviewed-on: https://chromium-review.googlesource.com/603874
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>