[autotest] Update shard state before master for modify_host
This is a testing-the-waters CL for the following idea: Any time there
is an rpc_fanout from the master AFE to the shards, the shard state must
be complete before the master state is modified.
This way, in case of a partial update, master state is left untouched.
Any corrective action assumes that the master state is correct in case
of master-shard desync, so that shard's partial updates will be
discarded.

Currently, we update master state before the shard in a few places. If
an update fails midway, this causes us to assume that the update was
successful, but the shards lag behind.

BUG=chromium:732999
TEST=modify hosts on master-shard on the staging servers.

Change-Id: Ia8920550f75e10e51f1c86eef347346c88e08e5e
Reviewed-on: https://chromium-review.googlesource.com/608580
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>