Improve autotest to run on non-servo equipped devices.
When running CFM tests on servos, we end up in a situation where
if we do not manually comment out a line in the control file, we
are not able to run tests.

This change introduces the possibility to specify an empty host
via --args "servo_host=". In this manner the user can explicitely
disable the servo host and run a test without it.

BUG=chromium:755512
TEST=manually tested on a panther.
Change-Id: If4de7809817c84cbb9160e59af43243bd9b2f032
Reviewed-on: https://chromium-review.googlesource.com/615245
Commit-Ready: Harpreet Grewal <harpreet@chromium.org>
Tested-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>