[autotest] Restore results_mocker code
So this is actually used.

In fact, Im quite surprised that nothing broke beside
lxc_functional_test.py.

That probably means that most of this code is unneeded.

Figuring out what exactly this does and what small subset of that is
actually needed is not something I want to do, so Ive partially
reverted the offending commit.

BUG=chromium:753267
TEST=Pre-CQ
TEST=CQ

Change-Id: I92d9e339ef23e3cd290767329c686c267853cadb
Reviewed-on: https://chromium-review.googlesource.com/616254
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>