CTS_N: skip ARM tests for novato-arc64
novato-arc64 is a x86 VM target and doesn't have any binary translators,
same as novato. So we have to skip the ARM tests for that board too.
No need to do the same for newbie-arc64 because there is no such board.

BUG=b:62840059
TEST=test_that cheets_CTS_N.arm.CtsDramTestCases against novato-arc64 no
longer fails (skipped).

Change-Id: I87292a70efe90e0d4989de4eb8d2da315edc0766
Reviewed-on: https://chromium-review.googlesource.com/617468
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>