cr50_test: fix original device image restoration
Only restore the device image if rootfs verification is disabled. If it
is not disabled, we can't do anything with the image at
/opt/google/firmware/cr50.bin.prod, but that also means the test did not
do anything. Restore the image if rootfs verification is disabled.
Always check that the version of the image is the same as it was at the
start of the test.

BUG=none
BRANCH=none
TEST=install new chromeos image, run firmware_Cr50BID, and verify rootfs
verification stays enabled. Run firmware_Cr50Update, verify the device
image is restored, and check that rootfs verification is disabled.

Change-Id: I88ec74824b0cb4896f283ef11701a72348659444
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/617597
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>